+++ tests/test_plugins/test_adjoint.py
@@ -35,7 +35,7 @@ from tidy3d.plugins.adjoint.components.data.dataset import JaxPermittivityDatase
 from tidy3d.plugins.adjoint.web import run, run_async
 from tidy3d.plugins.adjoint.web import run_local, run_async_local
 from tidy3d.plugins.adjoint.components.data.data_array import VALUE_FILTER_THRESHOLD
-from tidy3d.plugins.adjoint.utils.penalty import RadiusPenalty
+from tidy3d.plugins.adjoint.utils.penalty import RadiusPenalty, ErosionDilationPenalty
 from tidy3d.plugins.adjoint.utils.filter import ConicFilter, BinaryProjector, CircularFilter
 from tidy3d.web.api.container import BatchData
 import tidy3d.material_library as material_library
@@ -1498,6 +1498,17 @@ def test_adjoint_utils(strict_binarize):
 
     radius_penalty = RadiusPenalty(min_radius=0.2, wrap=True)
     _ = radius_penalty.evaluate(polyslab.vertices)
+
+    # erosion / dilation
+
+    image_01 = np.random.random(image.shape)
+
+    ed_penalty = ErosionDilationPenalty(
+        length_scale=1.0, beta=10, pixel_size=0.01, eta0=0.45, delta_eta=0.02
+    )
+
+    val = ed_penalty.evaluate(image_01)
+    assert val > 0
 
 
 @pytest.mark.parametrize(